Confirm delete (2): of edge on network pages and on conflicts table

+++ src/components/edges.tsx
@@ -133,7 +133,9 @@ function BaseEdgeWithDelete({
   }
 
   const deleteConnection = useCallback(() => {
-    onDelete(id);
+    if (window.confirm("Are you sure you want to delete this connection?")) {
+      onDelete(id);
+    }
   }, [id, onDelete]);
 
   return (
